requirements.txt
Flask
openai
requests
redis
pyyaml
python-dotenv